Fix Float negative-zero detection: zero-extend fromFlt bits
fromFlt returned the 32-bit float bit pattern sign-extended to i64, so -0.0
(0x8000_0000) became 0xFFFF_FFFF_8000_0000 and failed the Word comparison in
isNegZeroFloat32. That broke isNegativeZero and show's sign for -0.0 :: Float
(Double was unaffected, being full-width). Zero-extend instead. Fixes FArith,
RealFloat.

+++ rust/microhs-runtime/src/runtime/program/eval/dispatch.rs
@@ -259,7 +259,10 @@ impl Program {
             }
             "fromFlt" => {
                 let n = self.eval_float32(args[0])?;
-                self.int((n.to_bits() as i32) as i64)
+                // Raw float bits are an unsigned 32-bit word; zero-extend to i64 so the
+                // sign bit (e.g. -0.0 = 0x8000_0000) survives the Word comparison in
+                // isNegZeroFloat32. Sign-extending here breaks negative-zero detection.
+                self.int(i64::from(n.to_bits()))
             }
             _ => return Ok(None),
         };
